refactor: rename Hydrogen translation API
