+++ packages/hydrogen/src/utils/theme-text-store.ts
@@ -1,45 +1,1 @@
-/**
- * Minimal mutable store for design-mode static text overrides.
- *
- * useSyncExternalStore-compatible. Passed into ThemeTextProvider which
- * subscribes internally and merges overrides into merchantOverrides.
- * The studio bridge calls updateOverrides() via weaverse.internal.themeTextStore.
- *
- * Shape: flat Record<string, string> — dot-notation keys map to values
- * (e.g. "cart.title" → "Shopping Cart").
- */
-export class ThemeTextStore {
-  overrides: Record<string, string> = {}
-  private listeners: Set<() => void> = new Set()
-
-  setOverrides = (overrides: Record<string, string>) => {
-    this.overrides = { ...overrides }
-    this.emit()
-  }
-
-  updateOverrides = (newOverrides: Record<string, string>) => {
-    this.overrides = { ...this.overrides, ...newOverrides }
-    this.emit()
-  }
-
-  getSnapshot = () => this.overrides
-
-  getServerSnapshot = () => this.overrides
-
-  subscribe = (callback: () => void) => {
-    this.listeners.add(callback)
-    return () => {
-      this.listeners.delete(callback)
-    }
-  }
-
-  private emit = () => {
-    for (const listener of Array.from(this.listeners)) {
-      try {
-        listener()
-      } catch (error) {
-        console.error('ThemeTextStore: Listener error', error)
-      }
-    }
-  }
-}
+export { ThemeTextStore, TranslationStore } from './translation-store'
